Revisian 1

+++ database/migrations/2022_04_10_065135_create_penjualan_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('kode_penjualan');
             $table->integer('total_item');
             $table->integer('total_harga');
-            $table->tinyInteger('diskon')->default(0);  
+            $table->float('diskon')->default(0);  
             $table->integer('bayar')->default(0);  
             $table->integer('diterima')->default(0);  
             $table->integer('id_user');
